requirements.txt
requests
beautifulsoup4
trafilatura
markdownify
PyMuPDF
cloudscraper
selenium
webdriver-manager
Flask
Flask-SocketIO
gevent
python-socketio[client] 